1. update db script to remove the not null constraint on table requisition_item. 2. enhance the deployment script to support just restart tomcat.

+++ tomcatDeploy.bash
@@ -4,18 +4,26 @@
 # 1. set the variable of CATALINA_HOME
 # 2. set path to CATALIHA bin.
 #
-if [[ "$1" == "up" ]];then
-   git pull origin master
+if [[ "$1" != "restart" ];then
+   if [[ "$1" == "up" ]];then
+      git pull origin master
+   fi
+
+   if [[ "$1" != "nocompile" ]];then
+	   gradle -b build.war.gradle clean build
+   fi
 fi
 
-if [[ "$1" != "nocompile" ]];then
-	gradle -b build.war.gradle clean build
-fi
 shutdown.sh
 sleep 5s
-rm -rf ${CATALINA_HOME}/logs/*
-rm -rf ${CATALINA_HOME}/webapps/prm*
-cp build/libs/prm*.war ${CATALINA_HOME}/webapps/prm.war
+
+if [[ "$1" != "restart" ];then
+   rm -rf ${CATALINA_HOME}/logs/*
+   rm -rf ${CATALINA_HOME}/webapps/prm*
+   cp build/libs/prm*.war ${CATALINA_HOME}/webapps/prm.war
+fi
 startup.sh
-sleep 10s
+if [[ "$1" != "restart" ];then
+   sleep 10s
+fi
 tail -f ${CATALINA_HOME}/logs/catalina.out
